+++ src/components/Image/Image.stories.tsx
@@ -346,19 +346,22 @@ export const HTMLAttributes: Story = {
 
 export const FallbackImage: Story = {
   render: () => (
-    <VStack spacing="l">
+    <>
       <Heading
         id="fallback-heading"
         level="h4"
         size="heading6"
-        text="Fallback image"
+        text="Fallback Image Example"
       />
       <Image
         alt="Alt text"
-        src="foobar.jpg"
         fallbackSrc="https://loremflickr.com/2000/800/new+york+public+library"
+        onError={(_event) =>
+          console.error("Error loading image in Storybok example")
+        }
+        src="foobar.jpg"
       />
-    </VStack>
+    </>
   ),
 };